requirements.txt
flask>=3.0.0
pyserial>=3.5
numpy>=1.24.0
